api:  - minor code changes (files/structs moved inside package)
