+++ api/interface.go
@@ -1,0 +1,5 @@
+package api
+
+// ElrondProxyHandler interface defines methods that can be used from `elrondProxyFacade` context variable
+type ElrondProxyHandler interface {
+}
