+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/gin-contrib/cors v0.0.0-20190301062745-f9e10995c85a
 	github.com/gin-gonic/gin v1.3.0
 	github.com/pkg/errors v0.8.1
+	github.com/pkg/profile v1.3.0
 	github.com/stretchr/testify v1.3.0
 	github.com/urfave/cli v1.20.0
 	gopkg.in/go-playground/validator.v8 v8.18.2
